updated ui (create-notice-types)
